test(server): split oversized state tests
